+++ library/SolverUtils/UnsteadySystem.cpp
@@ -107,7 +107,12 @@ namespace Nektar
 
                 // Load generic input parameters
                 m_session->LoadParameter("IO_InfoSteps", m_infosteps, 0);
+                m_session->LoadParameter("IO_FiltersInfoSteps",
+                    m_filtersInfosteps, 10.0 * m_infosteps);
                 m_session->LoadParameter("CFL", m_cflSafetyFactor, 0.0);
+
+                // Time tolerance between filter update time and time integration
+                m_session->LoadParameter("FilterTimeWarning", m_filterTimeWarning, 1);
 
                 // Ensure that there is no conflict of parameters
                 if(m_cflSafetyFactor > 0.0)
@@ -137,9 +142,8 @@ namespace Nektar
             // Set up filters
             for (auto &x : m_session->GetFilters())
             {
-                m_filters.push_back(
-                    GetFilterFactory().CreateInstance(
-                        x.first, m_session, shared_from_this(), x.second));
+                m_filters.push_back(make_pair(x.first, GetFilterFactory().CreateInstance(
+                        x.first, m_session, shared_from_this(), x.second)));
             }
         }
         
@@ -245,17 +249,19 @@ namespace Nektar
             // Initialise filters
             for (auto &x : m_filters)
             {
-                x->Initialise(m_fields, m_time);
+                x.second->Initialise(m_fields, m_time);
             }
 
-            LibUtilities::Timer     timer;
-            bool      doCheckTime   = false;
-            int       step          = m_initialStep;
-            int       stepCounter   = 0;
-            NekDouble intTime       = 0.0;
-            NekDouble lastCheckTime = 0.0;
-            NekDouble cpuTime       = 0.0;
-            NekDouble elapsed       = 0.0;
+            LibUtilities::Timer         timer;
+            bool      doCheckTime       = false;
+            int       step              = m_initialStep;
+            int       stepCounter       = 0;
+            NekDouble intTime           = 0.0;
+            NekDouble lastCheckTime     = 0.0;
+            NekDouble cpuTime           = 0.0;
+            NekDouble cpuPrevious       = 0.0;
+            NekDouble elapsed           = 0.0;
+            NekDouble totFilterTime     = 0.0;
 
             while (step   < m_steps ||
                    m_time < m_fintime - NekConstants::kNekZeroTol)
@@ -312,6 +318,7 @@ namespace Nektar
                     ss << cpuTime << "s";
                     cout << " CPU Time: " << setw(8) << left
                          << ss.str() << endl;
+                    cpuPrevious = cpuTime;
                     cpuTime = 0.0;
                 }
 
@@ -368,11 +375,44 @@ namespace Nektar
                     ASSERTL0 (!nanFound,
                                 "NaN found during time integration.");
                 }
+
                 // Update filters
                 for (auto &x : m_filters)
                 {
-                    x->Update(m_fields, m_time);
+                    timer.Start();
+                    x.second->Update(m_fields, m_time);
+                    timer.Stop();
+                    elapsed = timer.TimePerTest(1);
+                    totFilterTime += elapsed;
+
+                    // Write out individual filter status information
+                    if(m_session->GetComm()->GetRank() == 0 && 
+                    !((step+1) % m_filtersInfosteps) && !m_filters.empty() && 
+                    m_session->DefinesCmdLineArgument("verbose"))
+                    {
+                        stringstream s0;
+                        s0 << x.first << ":";
+                        stringstream s1;
+                        s1 << elapsed << "s";
+                        stringstream s2;
+                        s2 << elapsed / cpuPrevious * 100 << "%";
+                        cout << "CPU time for filter " << setw(25) << left 
+                            << s0.str() << setw(12) << left << s1.str() << 
+                            endl << "\t Percentage of time integration:     "
+                             << setw(10) << left << s2.str() << endl;
+                    }
                 }
+
+                // Write out overall filter status information
+                if (m_session->GetComm()->GetRank() == 0 && 
+                    !((step+1) % m_filtersInfosteps) && !m_filters.empty())
+                 {
+                    stringstream ss;
+                    ss << totFilterTime << "s";
+                    cout << "Total filters CPU Time:\t\t\t     " << setw(10)
+                        << left << ss.str() << endl;
+                 }
+                totFilterTime = 0.0;
 
                 // Write out checkpoint files
                 if ((m_checksteps && !((step + 1) % m_checksteps)) ||
@@ -460,7 +500,7 @@ namespace Nektar
             // Finalise filters
             for (auto &x : m_filters)
             {
-                x->Finalise(m_fields, m_time);
+                x.second->Finalise(m_fields, m_time);
             }
             
             // Print for 1D problems