+++ CHANGELOG.md
@@ -81,6 +81,7 @@ v5.0.0
 - Fix TriExp.cpp orientation bug (!1048)
 - Fix XML attributes in conditions.cpp to be unordered (!1015)
 - Fix issue with HDF5 mesh input in serial (!1049)
+- Add estimate of filters CPU time (!1044)
 
 **NekMesh**:
 - Add feature to read basic 2D geo files as CAD (!731)